Move channel to top package (#29528)
