+++ kernel/data-pipeline/core/src/main/resources/META-INF/services/org.apache.shardingsphere.data.pipeline.core.channel.PipelineChannelCreator
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.data.pipeline.core.ingest.channel.memory.MemoryPipelineChannelCreator
+org.apache.shardingsphere.data.pipeline.core.channel.memory.MemoryPipelineChannelCreator
